Extract cbh_run from cargo-bench-history
Phase 8a of the cargo-bench-history fragmentation. Split the command run edge
out of the shell into its own implementation crate so this run/output layer is
mutation-testable in isolation:

- cbh_run: the RunOutcome a successful run returns and the RunError it fails
  with (aggregating ConfigError + StorageError), the OutputWriter port and its
  TokioOutputWriter/emit report writers, and the shared rebase path helper.
  Deps cbh_config, cbh_storage, cbh_diag, cbh_render (+ tokio, futures).

The in-memory MemoryOutputWriter fake is exposed across crate boundaries via a
private-test-util feature so the analyze orchestrators' tests keep driving it.
The shell re-exports RunError/RunOutcome/finish_with_flush from cbh_run, wiring
imports rebase from it, and the analyze modules import the output port from it.
