+++ Cargo.toml
@@ -34,6 +34,7 @@ cbh_git = { version = "=0.0.3", path = "packages/cbh_git", default-features = fa
 cbh_model = { version = "=0.0.3", path = "packages/cbh_model", default-features = false }
 cbh_probe = { version = "=0.0.3", path = "packages/cbh_probe", default-features = false }
 cbh_render = { version = "=0.0.3", path = "packages/cbh_render", default-features = false }
+cbh_run = { version = "=0.0.3", path = "packages/cbh_run", default-features = false }
 cbh_stats = { version = "=0.0.3", path = "packages/cbh_stats", default-features = false }
 cbh_storage = { version = "=0.0.3", path = "packages/cbh_storage", default-features = false }
 clap = { version = "4.6.1", default-features = false, features = ["default"] }
