+++ packages/cargo-bench-history/src/analyze/list.rs
@@ -19,6 +19,7 @@ use anyspawn::Spawner;
 use cbh_config::{Config, load_config};
 use cbh_diag::{Reporter, ReporterExt, StderrReporter, count_noun};
 use cbh_git::{GitHistory, SystemGitHistory};
+use cbh_run::{OutputSelection, OutputWriter, TokioOutputWriter, emit};
 use cbh_storage::{Storage, StorageFacade, resolve_storage};
 use jiff::Timestamp;
 use serde::Serialize;
@@ -30,7 +31,6 @@ use super::{
     facet_filtered_candidates, resolve_facets, resolve_now, select_dataset,
 };
 use crate::model::{BlessingRecord, DiscriminantSet};
-use crate::output::{OutputSelection, OutputWriter, TokioOutputWriter, emit};
 use crate::wiring::{
     cache_env, resolve_cache_path, resolve_config_path, resolve_local_path, resolve_project_id,
     resolve_repo, storage_env,
@@ -875,6 +875,7 @@ mod tests {
     use cbh_config::Config;
     use cbh_diag::RecordingReporter;
     use cbh_git::FakeGitHistory;
+    use cbh_run::MemoryOutputWriter;
     use cbh_storage::{MemoryStorage, Storage};
     use futures::executor::block_on;
     use jiff::Timestamp;
@@ -885,7 +886,6 @@ mod tests {
         BenchmarkId, BenchmarkIdPrefix, BenchmarkResult, EnvironmentInfo, GitInfo, Metric,
         MetricKind, Run, RunContext, ToolchainInfo,
     };
-    use crate::output::MemoryOutputWriter;
 
     fn config() -> Config {
         Config::default()
