+++ packages/cargo-bench-history/src/lib.rs
@@ -1,7 +1,6 @@
 #![cfg_attr(coverage_nightly, feature(coverage_attribute))]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![expect(
-    clippy::exhaustive_enums,
     clippy::exhaustive_structs,
     reason = "this crate's `pub` items form a handoff boundary between lib.rs and \
               the in-crate binary plus integration tests, used only inside this \
@@ -283,8 +282,6 @@ mod analyze;
 mod commands;
 mod config_writer;
 mod dispatch;
-mod outcome;
-mod output;
 mod wiring;
 
 pub(crate) use cargo_bench_history_core::model;
@@ -295,11 +292,11 @@ pub use cbh_command::{
     UnblessOptions,
 };
 pub use cbh_config::{ConfigError, default_template};
+pub(crate) use cbh_run::finish_with_flush;
+pub use cbh_run::{RunError, RunOutcome};
 pub use cbh_storage::{StorageError, StorageOverride, azure_backend_from_parts};
 pub use dispatch::{Overrides, run, run_with_overrides};
 pub use model::{
     BenchmarkId, BenchmarkIdPrefix, BenchmarkResult, EnvironmentInfo, EnvironmentProvider, GitInfo,
     Metric, MetricKind, Run, RunContext, SCHEMA_VERSION, ToolchainInfo,
 };
-pub(crate) use outcome::finish_with_flush;
-pub use outcome::{RunError, RunOutcome};
